Agent expects to add fdb entries for each pool member port that exists
Issues:
Fixes #754

Problem:
The agent will add forwarding database entries for each pool member
added to a pool. This means that it is expecting information from the
member's port to create those fdb entries. With the new change to ignore
port existence, we allow that member to get to the network_service.py
without a port. If no port is found, the member's provisioning status is
set to ERROR. Note that this is also the case when we are deleting a
member as well.

Analysis:
Removed the code in the update and delete member in network_service
module and put in a LOG.warning message to inform the user that no port
was associated with the member and no fdb entries were created/removed.

Tests:
Ran the test_member_status.py tests against this change. These were the
tests that failed and uncovered this issue.