Route domain cache lookup returns wrong route domain
Issues:
Fixes #768

Problem: Incorrect route domain ID is returned from the route domain cache.

Analysis: There is an error in the function that searches the cache,
so that when a cache entry is empty for a tenant, or the net name
is not found in a cached entry, the function returns an incorrect
route domain ID (the last route domain ID). Modified the function
get_route_domain_from_cache() to raise an exception when no entry
is found, then handle the exception in the calling function.

Tests: Added new unit test, test_network_service.