Pool will not be recreated if neutron-lbaas show it as 'ACTIVE' and pool does not exist on device
Issues:
Fixes #783

Problem:
When a pool is in an 'ACTIVE' or 'ERROR' state in neutron-lbaas, and the
pool somehow disappears from the BIG-IP device, the
_assure_pools_created method in lbaas_builder.py does not properly
recreate the pool because of faulty if/then logic. In fact, if 'ACTIVE'
or 'ERROR' is seen as the provision status, we will try to update the
pool instead of creating it. We should change this logic to better fit
the listener method.

Analysis:
Corrected the error and created unit tests to test the new
functionality.

Tests:
The new units tests pass. We will create a new tempest API test in the
driver to check this funcitonality as well.