Physical BIG-IPs supporting HPB should create their own ports on network.
Issues:
Fixes #825

Problem:
When a physical BIG-IP is connected to a SDN which supports HPB,
the presence of pool members which are mapped back from their subnet_id
to a Neutron network, should cause the BIG-IP to assure if an existing
port already exists and if not create a port on the appropriate Neutron
network. The port creation thus signals the SDN that segment connectivity
is required for the BIG-iP to complete the LBaaS request. There must be
an agreed upon way in the port attributes to signal the SDN that it is
the BIG-IP which requires a segment connection. With Cisco ACI the host_id
attribute on the port needs to be specified as the LBaaS agent_id.

This applies only to network connectivity required for pool members as
the load balancer side port is automatically created by the community
LBaaS plugin.

Analysis:
This commit modifies the is_service_connected() check to issue a port
create request to the driver if HPB is active and the network does not
have a segmenation ID.

This differs from the original solution in that it only creates one port
for the member network named bigip-opflex-<network uuid>, and this port
is created by the agent.

This solution does not create a port for a member if one does not already
exist.  The user needs to make sure that member ports are either created
by Nova when the virtual server is created or by the administrator, when
the IP address is for a server on an external network.  This will work if
there are no member ports; however, FDB entries will not be added for
members on the BIG-IP.

Tests:
Manual