Pool does not associate with listener if pool is created first
Issues:
Fixes #819

Problem:
When a pool is created and then the listener is created, the pool should
be linked to the listener via the _assure_pools_created method in
lbaas_builder.py. This does not work because the pool already exists on
the device and thus raises a 409 exception up to the function above in
lbaas_builder.py, causing the function to halt there. The listener is
never associated with the pool then.

Analysis:
Modified the _assure_pools_created call to only raise an exception if an
HTTP error seen other than a 409. If a 409 is seen, the call proceeds
and it is assumed the pool exists.

Tests:
Wrote new unit tests for this change. In addition to the system tests
that are currently failing in the driver, that should be enough to prove
this change is good.