neutronless loadbalancer test failure
Issues:
Fixes #693

Problem:
Multiple tests in neutronless/loadbalancer test suite fail against
liberty, mitaka, and newton with failure to find selfip object on
device.

Analysis:
Updated loabalancer tests to use the correct device name, which
masquerades as the hostname in the object naming schemes. We basically
reproduced the way the agent originally creates the object name to
create the expectation in the test.

Tests:
Ran tests in local sandbox.