Add smoke tests to agent's Makefile
Issues:
Fixes #841

Problem:
We need to add the option to make smoke tests for building on PR. We
should create a new make target to be able to run a set of specific
tests that we've picked as 'smoke tests'. These will run automatically
when a PR is created on this fork in Github.

Analysis:
Created a new target to run the smoke tests, and singled out a few tests
that should be run. These are all singlebigip tests.

Tests:
Ran those tests with the new smoke target against 12.1.1