smoke tests on PR do not fail when a test setup failure occurs
Issues:
Fixes #860

Problem:
The build on PR smoke tests, which are run when a PR is opened in the
agent, fail as expected when a test fails (tested in liberty), but the
build does not fail when the tests error out.

Analysis:
Removed the dash before the setup target in the Makefile.

Tests:
Running tests in build on PR.