neutronless esd test failure
Issues:
Fixes #692

Problem:
Running esd/test_esd.py fails against mitaka and newton stack with
AssertionError.

Analysis:
The tests were overriding an attribute of the lbaas_builder object,
which was causing the test failure seen in #692. This was not the
intention of the author. Once that was removed, the two failing tests
began passing.

Tests:
Ran full overcloud set of tests.