Agent manager lbdriver not initialized in proper order
Issues:
Fixes #893

Problem:
The lbaasv2 agent is not registering with neutron in the neutron agent-list upon
installation and start of the process. The service status looks as though it is active.

Analysis:
Moved the load_driver call to initialize the agent_manager driver attribute
before it is used to build configuration.

Tests:
Manual