update_fdb_entries in agent_manager.py does not expect proper data structure from neturon
Issues:
Fixes #782

Problem:
Okay, after doing some research into this issue, I've come to a few
points of clarification.

The only way I see an update_fdb_entries message kicked from neutron is
in the case that the member IP address has changed through a neutron
port update command.

In the case that the member IP address is changed via step 1 above, we
actually make no change on the BIG-IP device. An fdb entry does not
contain the IP address of the member.

If the member address changes in neutron, we must update neutron-lbaas
to that fact (this is not in the scope of this issue).

Analysis:
Instead of implementing code that would affect no change on the BIG-IP
device anyway, I simply commented out the call to update the fdb
entries and log a warning message when this event is seen in the agent,
thus showing the user a likely reason why the message was produced and
how they can remedy the situation.

Tests:
Wrote a new unit test to check the logging functionality.