Add device id to port creation for subnets RPC
Issues:
Fixes #963

Problem:
In order to support the new port binding scheme we are adding the
loadbalancer id as the device_id for any ports that are created as
part of a loadbalancing service. This include self IPs and SNATS.

Analysis:
Extending the create_port_on_subnet() RPC call to accept 2 new
parameters: device_id and binding_profile.  This change required
a modification to the calls to create the SelfIP and SNATs that
added the loadbalancer ID as a parameter to these methods.  The load
balancer ID is used as the device_id of the Neutron port.

Tests:
tests/functional/neutronless/loadbalancer