Misconfigured LOG var results in AttributeError: 'function' object has no attribute 'error'
Issues:
Fixes #685

Problem:
Exception occurs when trying to log from module that has not been
properly instantiated with a name.

Analysis:
Used @busterswt solution and wrote a test to check the logger.

Tests:
A new unit test was added to check that the logger is named what we
expect and that it has callabled functions such as error.