Fixes Mishandle of the VS's getting the LBID as addr
Issues:
Fixes #985

Problem:
* LB that's pointed at by VS as a dest may not be handled properly
  * Address is set to LB's name/id
  * Code is expecting an IP address, not a name string

Analysis:
* Loads the address from the associated VA
  * Uses the VA if this is possible
* When a valid address is used, an exception does not keep the loaded
  * Instead, the valid address is used
* If nothing of this is recognized as expected values, except rise

Tests:
This is a cherry-pick from mitaka that fixes a known bug with the shared
networking features on both liberty and mitaka.  This test is now passing in
mitaka, and the second set of tests are unit tests and are white-box tests.
These are introduced via `test_network_helper.py` as `network_helper.py` was
the impacted module of code for this fix.