Issues: Fixes #1030
Problem:
The vcmp tests have strayed in various ways, and since these tests are
not running nightly, we need to get them back in shape. This includes
running them to ensure no functionality was broken. If anything is
found, we'll file additional bugs and fix them.

Analysis:
Fixed up a few changes made in the way these tests setup and one change
in the expectations they make about the disconnected networks.

Tests:
Ran all tests and they pass. The results will be posted in the PR.