update requirements/code on master
The code on the master branch was left in the state
applied by merging forward stable/newton, however
master should refer to "master" in other repos, not
to release branches.

Additional, minor changes were applied to rectify
obvious import errors, and style issues.

Note: some of the import errors are caused by
changes in the interfaces exposed by the
underlying libraries.  This is expected as the
master branch depends on the Queens version of the
community code.

* fix some neutron.common.exceptions errors
* add pytest-symbols to requirements.functest.txt
* make an import relative
* fix flake8 style issues