Remove device interface information from configuration
Issues:
Fixes #1038

Problem:
Including the interface information in the agent configuration
could possibly exceed the amount of space allocated for in the DB.

Analysis:
Since this was added for the purpose of determining whether or not
an agent is controlling VE instances managed overcloud.

Add an attribute to the configuration that will indicate whether the
BIG-IPs are VE instances deployed in Nova.

Tests:
TBI