Implement device management robustness changes.
Fixes:
Issue #730

Problem:
When an agent cannot communicate with BIG-IPs we cannot kill the agent.
Instead, we need to report another configuration value to report the operational
status of the BIG-IP device group.  If the agent cannot communicate with any of
the backend BIG-IPs this change allows the agent to communicate that it is DOWN

Analysis:
This change allows for the dynamic entry/exit of BIG-IP from the device group
that an agent manages.  It implements the logic to detect when the BIG-IP is no
longer available.  When the set of managed BIG-IPs reaches zero, the agent will
advertise itself as administravtely down.  The driver can then perform rebinding
of loadbalancers to another agent in the same environment.

Test:
Tested in POC and unit tested.