lbb test raises unexpected swallowed TypeError
Issues:
Fixes #1073

Problem:  The test was not testing the behavior we expected it to.

Analysis:  A mock used in the test was generating a TypeError which was
swallowed by a Scantily-Clad except statement:  i.e.

"except Except"

The Scantily-Clad Exception caused the test author to be confused about the
correct behavior of the agent.

Tests:  N/A