This commit  has a few minor fixups to ease test.
The changes in test_esd remove the use of the
'Iterator.next()' in favor of the simpler pattern
of indexing into the list of services.

The practical effect is that test developers can now
remove sections of test code without changing the
meaning of the "service" variable in subsequent
portions of the code.

The only other change is dedentation of a few lines
to shut flake8 up.

The third commit (now squashed) built on the previous
two to refactor the tests such that they have conventional
setup and teardown functionality. These setups and
teardowns are not the ideal of booting snapshotted vms,
but they do modularize the tests and speed development in the short term.

ORIGNAL COMMIT MESSAGES BELOW:
fix flake8 violations

simplify loading of service replays in test

refactor tests to be more modular to speed test development