Pool/healthmonitor/member changes for lbaas_builder reorder
Problem:
The pool no longer needs to update a listener because we are moving the
virtual create/update to after the pool creation and the virtual delete
to before pool creation.

This code removes a lot of the look back logic to find the listeners
that are associated with this pool.

Also the monitors are now being created before the pools so the look
back logic from monitor to pool was removed.