Reimplement listener deploy for new lbaas builder ordering
Problem:
Since the order of operations was changed, the listener now needs
to configure it's default pool itself.  It also needs to configurare
any associated persistence profiles that are defined by the pool.

Some of the logic to create a virtual server definition was consolidated
from the listener service and lbaas_builder to the service adapter.
An entire vip needs to be created as a part of the call to get_virtual
in the lbaas_builder.  This way we have a complete definition of the
VIP before we make any calls to the BIG-IPs