Reimplement L7 policy for lbaas_builder reorder
Problem:
The L7 policy code was written in a way to support the modification
of a predeployed virtual.  Since we are now deferring the creation
and update of virtuals to after the assure operations are performed on
all its dependencies, we can remove much of the logic to configure
the listeners to the assure_listener.

Removed files that are not needed for this refactor and create the
entire policy in assure_l7policies_created.  This step aggregates all
the policies defined in the service to the associated listeners and
then deploys the policy on a per-listener basis.