iControl driver bigip attributes not properly initialized
Problem:
If the driver is in f5_global_routed_mode, then the local_ip attribute of
the bigip will not be initialized.  This causes an issue with agent
configuration reporting.

When enumerating device interfaces as part of post initialization we might
iterate over bigip instances that have performed initialization.

Analysis:
Check for the presence of the local_ip attribute.
Initialize the device_interfaces attribute with an empty dict in
early initialization.

Test:
Run testing in Global Routed Mode
Run testing in multi-agent configuration.