Fix listener_delete and healthmonitor_create assure methods
# Using the Fixes #<issueid> will close the issue on commit to repo
Issues:
Fixes #

# Describe the issue that this change addresses
Problem:
From the review and testing it was apparent that the listener teardown was
still relying on the old method of throwing exceptions from the assure.

The healthemonitor assure created method was not properly setting the status
of the service object.

# Describe the change itself and why you made the changes you did
Analysis:
Implent the assure_listeners_deleted to handle a return value and
set the provisioning status based on that.  Remove the exception raise.

Change the healthmonitor assure create method set status to active only
when no provisioning errors have occurred.

# Describe the tests you ran and/or created to test this change
Tests:
test/test_lbaas_builder.py