Pool deployment refactor
Problem:
The update of a pool in assure_pools_created causes the pool to lose member
configuration until the members are created in the member assure method.

Analysis:
Change the implementation to configure the members on the pool at the time
the pool us updated/created/deleted.

Add a new method to assure the pool members that simply checks for the
existence of the members.  The update_opeating_status method could be
called here to get the immediate operating status of pool members.

Added a node cleanup to the pool delete and assure pool member created.

Test:
test/test_lbaas_builder.py
test/test_pool_services.py