Add demo_policy fixture for ESD tests
Problem:
Some ESD tests attempt to configure a virtual with a non-existent policy 'demo-policy'.
For these tests to pass, there must be a demo-policy deployed on the system.

Solution:
Added fixture demo_policy that creates a policy, with a rule and publishes it.
This fixture is passed to those ESD tests that require demo_policy.