refactor neutronless test infrastructure
Issues:
Fixes:  This is a test issue.

Problem: In the previous iteration of the neutronless test architecture all
configuration was accomplished with a tall stack of fixtures.  This worked for
a given set of tests, but proved to be very difficult to modify.

Analysis:  In order to make it easier to write different kinds of tests
the bulk of configuration has been migrated into a standard Python class.  This
class is instantiated within a fixture, guaranteeing that entry into the test
system still conforms to the behaviors enforced by pytest.

The result is a much more flexible system that maintains the feature set of
pytest.

Tests:  I ran the tests inside neutronless/esd