Add test for multiple esds are applied and removed
Problem:
Add a test to verify when mutiple ESDs are applied and one of the ESD profile is removed the
other still remains.

Solution:
Add necessary testing for it in test_esd_pools.py which applies 2 ESDs, deletes 1 and checks
if the other ESD is still present. This test makes sure issue #1044 is fixed.