Listener service fails to create SSL client profile (#1170)
Issues:
Fixes #1164

Problem:
Creating a TERMINATED_HTTPS listener fails and results in a
service in an error state.

Analysis:
As part of the code refactor much of the implementation of the SSL
creation/addition was retained from the previous implementation; however,
the real vip object was not updated by the add_ssl_profile function.

The dummy vip has no profiles element and so we get a key error
when we try to configure the attribute.  This changes passes the vip
configuration object created by map_virtual into the add_ssl_profile
function.

Tests:
functional testing for listener.