SSL client profile not created on backup BIG-IP
Issues:
Fixes #1201

Problem:
When trying to deploy a terminated HTTPs listener, the loadbalancer goes
to error state. This problem only occurs when there is more than on BIG-IP
configured.

Analysis:
The problem is due to way in which SSL profiles are created on the BIG-IP.
As we iterate over the BIG-IPs in the DSC we create a client SSL profile on
each BIG-IP. We also modify the icontrol REST payload that describes the
virtual server. The problem occurs because we are adding a ssl client
profile twice to the VIP object.

Tests:
test_listeners_tls
test_listeners_tls_sni