Can't create loadbalancer in Global Routed Mode
Issues:
Fixes #1215

Problem:
In an overcloud test to create a loadbalancer, the driver throws
an exception because the advertised tunnel type was initialized
to 'None' in the agent configuration.  If there is no local VTEP
ip address, this advertised tunnel type should be initialized to
an empty list.

Analysis:
Changed initialization of tunnel_type from 'None' to an empty list.

Tests:
Overcloud testing.