Issues: Fixes #1261
Problem:
There are neutron imports throughout the agent code and we need to move them into one file.

Analysis:
This change moves all the neutron constants to F5 agent constants file. Also it moves the neutron exceptions to F5 agent exception file.
The changes are made in both unit tests as well as bigip/ directory.

Tests:
To test the changes, I started my local agent on my stack and ran the community api and scenario tests as well as systest-common api and scenario tests. I also ran the agent functional tests against a bigip. All of the results are consistent with the trtl results.