Bugfix: handling ReferenceError
Issues:
There exists a race condition that could result where a fdb gets called
interrupting things after a tunnel has been destroyed before the
tunnel's weakref gets cleared.  While there is no logistical way of
clearing this cleanly, it will result in a ReferenceError.

This change silences this weakreference failure.  At this point, it does
not matter if an FDB entry comes in on a tunnel that is being destroyed.
This includes members' vteps that come in post-destruction of the
tunnel.

This is a standard side effect of having to lock caches and a waiting
child never gets the chance to finish.

Problem:
* ReferenceError was thrown due to a destroyed tunnel before the
member's vteps are removed via deletion cycle of a Tunnel

Analysis:
* Uses a decorator to silence the ReferenceError during member deletion
  * Would only happen if a tunnel was destoryed BEFORE member teardown

Tests:
I'm not certain that there's a threat to tests; however, in my
estimation, the tunnel was acquired by the ReferenceError child, then it
released its lock, and the periodic clean-up task would come in and
potentially destroy the tunnel.  The ReferenceError child simply did not
have enough time to perform its cleanup in time.

Now, there does exist a risk that the ReferenceError was thrown before
member and other objects were destroyed appropriately on the BIG-IP.
This would simply mean that we do not have sufficient decorator
coverage handling the ReferenceError raises.