Adding additional logging for efficiency
Issues:
The post-networking side of icontrol_driver's _common_service_handler
jumped to 15 seconds.  This change will attempt to capture how long it
takes for each process to attain a lock and how long it keeps that lock.
Further, an additional time delta provides how long a each method takes.

Problem:
* Post-networking algorithm took a 15 second spike

Analysis:
* This grabs 2 deltas one for long methods take and the other for locks
  * deltas are captured just before each during the cache.lock decorator

Tests:
This is to be used with tests