Adding HTTPError handling for remove_multipoint
Issues:
Virtuals were found to be lingering in tests while
remove_multipoint_tunnel was executed.  This handler will cause the NCH
to continue to track the tunnel, but also log the redundant attempt.

Problem:
* ICD will execute code that will attempt to del tunnel before the VS

Analysis:
* This results in a 400 error for a dependency conflict on big-ip

Tests:
Our tests ran into this.  Note that this is a patch of sorts and should
be corrected appropriately at the VS level as well to prevent other
types of potential conflicts.