Tagmode incorrectly set when creating VLANs
Issues:
Fixes #1291

Problem: The agent always sets tagMode attribute to "services" value.
This causes a failure when running against BIG-IP version 13.0.0.

Analysis: The tagMode attribute only applies when configuring BIG-IPs
for QinQ VLANs. This is not the case for OpenStack. Version 12.1.2
ignores tagMode, but version 12.1.3 and 13.0.0 validate the attrbiute
and require that it not be set. The fix is to remove setting tagMode.

Tests:
Functional tests test_loadbalancer_disconnected:
  test_featureon_nosegid_to_segid_lb
  test_create_delete_basic_lb
  test_featureoff_create_delete_basic_lb

Unit tests::
  test_vlan.py