patch for sdn-base
For differences between sdn-base and 9.8-stable branch(commit 46e2320).

Note: version in __init__.py is modified to be 5.0.0 here.
So it is actually not the exact differences between the two branches.
Meanwhile, it only includes the files from f5_openstack_agent directory.
Other directories are ignored here for now.

NOTE: This should never be merged into sdn-base branch.
The sdn-base branch should never be touched/merged.
sdn-base is at 14d8557.

This should only server as the PR to download the patch file.