fix: bigip device down, the whole process exit
When bigip is down the whole F5 agent process will be exit.

This is the icontrol sdk sent timeout signal to the process when it is
connecting to the bigip device, but on one will reset the signal.

refer to sdk code: f5_sdk-3.0.11-py2.7.egg/f5/bigip/__init__.py