[OPENSTACK-2282]Fix snat creation conflict issue.
This issue happens in multiple agents environment.
When there are 2 or more loadbalancers are created at the same time,
the 2 loadbalancer creation logic from different agents may run into
the situation that both of them want to create the resources (of the same name).

This issue is similiar with route domain creation conflict described in PR 1654.
But the solution is different.