OPENSTACK-2479: dual-stack-snat
one network only has 1 or 2 network.
each subnet will have its snat ips in snat pool.

when delete, we check vlan is used or not,
if it is used, it means selfip, vlan, route domain can not delete,
because we need theset shared network resource to do L3 route for
different subnets in the vlan network.