OPENSTACK-2799: fix to delete pool
if bigip partition is missing,
we can not delete pool in neutron db.
it is because pool has relationship with pool member and listener.

when the partition is missing,
member can not update(create if missing)
allocate route domain in the misssing partition.

when the partition is missing,
listener can not update(create if missing)
in the missing partition.